qa/requirements.txt
pytest-playwright>=0.5.0
pytest>=8.0.0
